Introduced new options for .ccls files
in order to target some rare corner cases in project organization.

%inferhint <relative path hint>
this option can be used for a case, when source
 files being edited, viewed are not actually used
  to build the target, but they are in some
automatic way included in other auto-generated
files (like jumbo builds), so the standard infering algorithm might
not find the proper passing target with options
this 'hint path' will help it

%fake_* (must be combined with %cpp or/and %h, etc.)
make some/all files 'fake', meaning they should not be used as direct targets
when building a compiler invocation, but instead
an 'inferred' file will be 'compiled' (for code completion, diagnostics,
 parsing)
this allows to work normally with e.g. header files, that are
not self-contained (have no includes on their own)
relying completely on them being included in some
sort of a 'master' header (like stdafx PCH one) in the right order.
In this case it makes sense to 'compile' the some original TU, but work
the 'magic' (perform code completion request) on the original one.